requirements.txt
streamlit==1.31.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
pandas==2.1.0
python-dateutil==2.8.2
nltk==3.8.1
tweepy==4.14.0
praw==7.7.1
pytrends==4.9.2
markdownify==0.11.6
